tests/pyproject.toml
[tool.poetry]
name = "hr-manager-tests"
version = "0.1.0"
description = "End-to-end tests for HR Manager"
authors = ["Dimitri Merejkowsky <dimitri@dmerej.info>"]

[tool.poetry.dependencies]
python = "^3.9"
playwright = "^1.17.2"
pytest-playwright = "^0.2.2"
Faker = "^10.0.0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
invoke = "^1.5.0"

black = "^21.12b0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
